Initial commit
Using just SDL right now, no physics.
Using a weird threadDelay hack to get a timer, which means I have to
use SDL.waitEvent and not SDL.waitEventBlocking
Draws any row-based character animation from
<http://lpc.opengameart.org/static/lpc-style-guide/assets.html>
in a loop that is not tied to frame rate.

Closing the window quits the program.