Chipmunk/Hipmunk Physics
Hack in some physics.  I'm using the top-down joint hack from the Chipmunk
Tank demo.  In order to get the MaxForce and MaxBias properties I had to
add them to Hipmunk, so this will not build against upstream.  I will likely
have to publish my fork or bundle it until upstream accepts my patch.

The sprite now walks under the control of chipmunk.  There is an invisible
line it walks into in order to test that collisions are working.  They are.