Initial player control
Hard-coded arrow keys for the player.  Hard-coded speed.
Uses rotation of a speed vector to ensure that speed is constant
in any direction.
Only change the component of the direction that is set by the key
in question, so that roll-over works nicely.
Every single arrow key down handler calls pgo to set the speed > 0,
this could probably be done more nicely.