Configurable player animations
Parser for a player config file containing the specifications for all of the
animations for a player (one animation per direction).

Actually store all of these animation specs in the player, and source the
initial animation from among them.

Need to support advanceAnimation for single frame animations, so that was
also added here.