Player and controls selection
Start off in a seperate screen that allows players to join by selecting
their controls.  Players can then use their selected E/W controls to
choose a player character.  Players can use their selected Start control
to start the game.

Once the game is started, there is still only one player actually shown
and controlled, but obviously this needs to change.

Using the PortLligatSans font for the text on the selection screen.

The text on the selection screen currently uses the internal name for
the action you are binding.  This needs to change to a nice name.

The code for this whole thing is a bit dense, and very do-notation
heavy, much like the rest of this hackfest.  More cleanup would be good.

Controls are now stored in the player structure, which seems a
reasonable place to put them.