Basic attack framework
Single-action attacks roughed in.  Basic forward-looking stuff for charge-up
and for projectiles is in.

Every attack is modeled as a projectile, and some projectiles just have a very
short life.

Need to change current press-key-to-interrupt model to a two-stage queue model.

Question: if the frame rate drops, and many timer events are in the SDL queue,
will arbitrary shapes potentially get deleted?  On the one hand: probably.  On
the other hand, if they get processed fast enough, probably not.  Needs better.

Damage is being stored and printed out to STDOUT just now.  Needs better.

Damage is not dealt until the "projectile" hits the target.  This happens
*after* the release animation has fully played out.  Seems reasonable.

All of this will need serious cleanup.  I just want it to work.