Better queuing of actions
Hold an action in reserve for smooth button spamming.
Replace the held action on new button press, but do not move the held action
into current until current action is complete.
If there is no held action, then there will be no action.
Only update animation (which resets to first frame) if something actually
changed.  This actually improves even existing animation look.
Holding button is still "charge up" even for 0-length charge-up, which looks
funny.  Needs better.
Held action gets counted as starting the moment the current action ends, and,
if button was already released, ending (now + (realEnd - realStart)).  This
makes it possible to end in the future, which messes with unsigned Word32 math,
so the "is it done yet" math is now all done toInteger.

handleAction is now actually a partial function in terms of the states of the
abilities.  This is because some states should be impossible and would indicate
a bug, so the partiality is equivalent to an assertion in this case.  When we
fix all warnings we will probable want to put a sane (error "") call in.