Seperate charge and release animations
Improve animation model and file format to seperate charge-up and release
animations for abilities.  So there are now two kinds of animation:
SimpleAnimation which loop over a single path (one per direction) and
AbilityAnimation which store the ability in question and a set of animations,
one per AbilityState (charge and release, and then sub stuff per direction
still).

Frame rate for charge-up and release animations are set by the chargeLen and
releaseLen, allowing a slow charge up with a quick release.

AbilityAnimation do not loop, but rather hold on their last frame.

The start frame (col) has to be kept now, so that frame rate can be accurately
determined even on animations that start partway into the animation and then
stop.  This is still sort of simplistic, but I did it so that charge/release
can be in the same row, since this is common.