Display projectile animations
Actually draw projectiles.  Projectiles need to move, so velocity is added.
Projectiles need to keep animating after hitting (need to end on last frame),
and really we want the velocity to keep going during that since it manifests as
part of the animation, so we just set a deathPos thing which mostly acts as a
flag to take it out of the collision possibilities, but it doesn't get removed
until the animation next hits the last frame.  This is a bit messy, but works.

Also, extend the parser to allow projectile animations to be specified.  If
none is specified, there is none and no projectile will be drawn.  Projectile
spec also contains the life of the projectile and the framerate of the
animation.

Projectile position is a big odd, because the "front" of a projectile is
different depending on orientation (determined by which way it is moving) but
we always draw from the top-left corner of the sprite.  The current position
function should cause the sprite to be drawn such that the physics object is
right at the "front".  That is, the bottom for E,S,W and the top for N, the RHS
for E and the LHS for W, the middle for N and S.