Basic knockback which cancels action
Also roughs in a basic block concept, though we're only using it on fall
right now.

While blocking (falling) you travel at a fixed velocity: probably 0 in
many cases.  In this case it's set by the knockback of the attack that hit
you.

Blocking (falling) is an action.  It clears your current action and queued
action, so that charge-up gets interrupted by any attack.

Speed is also set to 0 so that you stop when hit.  This may not be desirable,
may need tweaking.

Also, knockback starts right away, which can look funny for long animations
(such as fire lion).