Big ugly commit for energy
I sort of rabbit-holed while writing this one.

I saw repeated code for generating random locations, getting item
positions, and drawing items, and decided to refactor it to be a bit
cleaner.

Biggest change: there is now a class Drawable which you can ask for
its position, draw, advance state (usually animations), and draw
a list of them sorted by y (for isometric Z)

Upshot in the end: energy pellets are randomly dropped and can be picked
up for +10 energy.