Support Music and Sound Effects
Music plays in the menu, then switches to the music for player 1 once the
game starts.  After that, music switches to whatever player currently has
the least deaths.

Sound effects can be specified on each action, and the sound will be played
when the action is performed.

Be careful in the code: if the runtime detects no references to a music or
sound effect, it will garbage collect them, and they will stop playing if
currently playing.  Just being in scope is not enough to keep a reference,
it has to be in scope and be used somewhere.

This is really only a problem for menu music right now, and is the reason the
menu music gets passed to startGame.