Players have colours
Use an HSV-based system to get a unique colour for each player.
Bind to SDL_gfx (not using the bindings because they expect colours in
a dumb, hard-to-get way) to draw circles in these colours under players' feet.

This circle is in fact the physics collision circle.  Now that we can see it
the down-below bias became even more apparent.  Move it up a bit.