Refactor event system
It turns out, SDL just does a poll-and-wait internally.  Not just the
Haskell binding, but actual backend SDL.  What?  So, ok, we can do that
ourselves better.  We can provide a polling loop that times out after
some time (like select() would) and use that to build our framerate
driver instead of using a second thread.

Given the way that threadDelay interacts with FFI calls on the same
haskell thread, this now needs to be compiled with -threaded or else it
can experience lock-ups.  I don't fully understand why, but it works.

IDLE time is now tracked by the profiler too, which is nice.  >90% IDLE
on my system!

While we're at it, refactor the "next" wrapper to do the actual event
poll and such, thus getting us out of the `do e <- wait; case e of`
synantactic anti-pattern.  Yay!